Dodanie komponentu przycisków oraz naprawienie back button
